Initial Commit for i18n-sonder
This commit contains basic scaffolding for this gem,
and defines the DefaultLocaleOptimizedKeyValue backend.

All the scaffolding is auto-generated via bundle gem.
The custom backend is copied directly from
Flatbook/flatbook#10560.

This backend is derived from Mobility's default KeyValue
backend, used to store translations of model text/strings,
except that when the locale is the default
i18n.default_locale it directly queries the model's tables
instead of querying the translation tables.
This customization optimizes read/write performance for
the default locale.