Create new CrowdIn Adapter for every worker perform call
When the Clockwork gem executes a worker's perform() function, it does
not always create a new worker. As a result, we end up using the same
localization_provider every run. This means that all the caching is
preserved across multiple runs, which is not the desired behavior. We
want a brand new client object every run. This PR makes that change.