Only upload new translations if attribute value changes
Previously, we were uploading content for translation regardless of
whether the content was different or not. This leads to quite a large
amount of API requests made to CrowdIn, which is wasteful. Here, we make
the change such that only if the value is changing do we schedule a job
to upload content for translation.