Find duplicate translations before uploading for translation
This is an optimization performed asynchronously in the UploadSourceStringsWorker.
We will now look through the translations tables, and look for existing
translations for the strings we are planning to upload for translation.
We apply the duplicate translations, and if all locales have existing
translations for any attribute, we do not upload that attribute for
translation.

CAVEAT: This optimization assumes that the Key-Value Mobility backend is
being used. It is not implemented in a flexible enough way to be
backend-agnostic.