Allow to namespace source strings in CrowdIn into directories
This introduces the concept of namespacing source strings. Models that
use the UploadForTranslation Mobiliy plugin can now implement a
`namespace_for_translation` method, which returns an array specifying
the namespace hierarchy to store source strings in.

In CrowdIn, this namespace hierarchy will be organized as directories.

--- CAVEAT ---
Even though the namespace is represented as an array in order to
represent a hierarchy, in this PR we DO NOT allow for nested directories
in CrowdIn. This is eventually desired, but can be done as a follow-up.